pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-ally"
version = "0.1.0"
description = "Meta agent for creating and configuring AI copilots"
readme = "ReadMe.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.15.0",
    "datamodel-code-generator>=0.35.0",
    "inform-ai-core[authorization]==1.13.0",
    "ipykernel>=7.0.1",
    "openai>=1.35.0",
    "pydantic-ai>=1.2.0",
    "pydantic-ai-slim[ag-ui]>=1.2.0",
    "pydantic-evals[logfire]>=1.2.0",
    "pytest>=8.4.2",
    "ruff>=0.14.8",
    "trio>=0.31.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["meta_ally*", "examples"]
namespaces = false

[tool.uv]
index-strategy = "first-index"

[[tool.uv.index]]
name = "INFORM_AI"
url = "https://inform-ai-637423203277.d.codeartifact.eu-west-1.amazonaws.com/pypi/inform-ai/simple/"

